# parser: Optimize mem alloc in parser