<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
The problem, that this pr aims to solve, is too much pressure on gc when it comes to long query(more than 5000 characters).

### What is changed and how it works?
The way to solve the problem is created and released objects from `sync.Pool`. 

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test

Code changes

 - Has exported function/method change


 - Has persistent data change

Side effects

Benchmark is as follows:


```

benchmark                   old ns/op     new ns/op     delta

BenchmarkParseComplex-8     256011        219986        -14.07%

 

benchmark                   old allocs     new allocs     delta

BenchmarkParseComplex-8     761            666            -12.48%

 

benchmark                   old bytes     new bytes     delta

BenchmarkParseComplex-8     72398         71639         -1.05%
```